frameparams: port to zstd 1.3.0
ZSTD_frameParams was essentially renamed to ZSTD_frameHeader. The
type of one of its members was also changed.

We should probably rename the python-zstandard API. But first thing's
first: let's get python-zstandard to compile with 1.3.0.

+++ c-ext/python-zstandard.h
@@ -49,7 +49,7 @@ extern PyTypeObject CompressionParametersType;
 typedef struct {
 	PyObject_HEAD
 	unsigned long long frameContentSize;
-	unsigned windowSize;
+	size_t windowSize;
 	unsigned dictID;
 	char checksumFlag;
 } FrameParametersObject;
